+++ libraries/reconstruction/ml_align2d.cpp
@@ -443,7 +443,7 @@ void ProgML2D::produceSideInfo2()
     if (model.do_norm)
     {
         average_scale = 1.;
-        for (size_t refno = 0; refno < model.n_ref; refno++)
+        for (int refno = 0; refno < model.n_ref; refno++)
         {
             model.scale.push_back(1.);
         }
@@ -640,7 +640,7 @@ void ProgML2D::preselectLimitedDirections(double &phi, double &theta)
     pdf_directions.clear();
     pdf_directions.resize(model.n_ref);
 
-    for (size_t refno = 0; refno < model.n_ref; refno++)
+    for (int refno = 0; refno < model.n_ref; refno++)
     {
         if (!limit_rot || (phi == -999. && theta == -999.))
             pdf_directions[refno] = 1.;
@@ -1969,7 +1969,7 @@ void ProgML2D::maximizeModel(ModelML2D &local_model)
         ASSIGN(wsum_sigma_noise);
 
 
-    for (size_t refno = 0; refno < local_model.n_ref; refno++)
+    for (int refno = 0; refno < local_model.n_ref; refno++)
     {
         double weight = sumw[refno];
         if (weight > 0.)
